Changed approach
Request the product details in the popup.html itself through the content endpoint on the front-end.

The is original approach (parse html or request the content endpoint) on the content.js file. This code has been commented out and not removed because there is potential use for it still. If we want to get the response cached version for the product details (to compare cached vs uncached) we need to parse the dom and extract the page data json and send it back to the exension popup.html